Modify the logic inside the useEffect.
Using the state called hasError, when an error is received in response to an Api request, modify it to raise the Toast alarm only once, instead of requesting it infinitely again.

Change the State to Ref in fetch useEffect.

Now, at the time of the search action or the first page loading, fetching is requested only once.

# Additional
- Remove document-collection name-by sorting function in Usage Statistics Management. Not supported by backend api.
- Modify the default sort value in usePagination.js to prevent errors when there is no Id fields in data. So, every component which using usePagination.js is effected.